+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 ## Unreleased
 
+### New capabilities
+
+- Made `/workflow` run and agent actions discoverable through Enter while retaining the `a`/`A` shortcuts ([#185](https://github.com/vekexasia/pi-extensible-workflows/issues/185)).
+
 ## [5.1.0] - 2026-07-31
 
 ### New capabilities